# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
      "title": "Symfony API REST",
      "description": "CRUD API in Symfony",
      "version": "1.0.0"
  },
  "paths": {
...
      "/api/products/{product}/get": {
          "get": {
              "operationId": "get_api_productsget_product_by_id",
              "parameters": [
                  {
                      "name": "product",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
      "/api/products/create": {
          "post": {
              "operationId": "post_api_productscreate_product",
              "requestBody": {
                  "required": true,
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/ProductDto"
                          }
                      }
                  }
              },
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
      "/api/products/{id}/update": {
          "put": {
              "summary": "Update a product by the given ID. Requires full product data in the request body.",
              "operationId": "put_api_productsupdate_product",
              "parameters": [
                  {
                      "name": "id",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "required": true,
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/ProductDto"
                          }
                      }
                  }
              },
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
      "/api/products/{product}/delete": {
          "delete": {
              "summary": "Delete a product by the given ID. Validates the authorization token and performs the deletion.",
              "operationId": "delete_api_productsdelete_product",
              "parameters": [
                  {
                      "name": "product",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
...
          "ProductDto": {
              "required": [
                  "name"
              ],
              "properties": {
                  "name": {
                      "type": "string"
                  },
                  "description": {
                      "type": "string",
                      "default": "",
                      "nullable": true
                  },
                  "price": {
                      "type": "number",
                      "format": "float",
                      "default": 0,
                      "nullable": true
                  },
                  "categoryId": {
                      "type": "integer",
                      "default": null,
                      "nullable": true
                  }
              },
              "type": "object"
          }
      }
  }
}